+++ src/core/router/CHANGELOG.md
@@ -13,7 +13,7 @@ Changelog
 
 #### :bug: Bug Fix
 
-* Fixed alias priority, when resolve path parameters
+* Fixed alias priority when resolve path parameters
 
 ## v3.50.0 (2023-06-??)